requirements.txt
groq
python-dotenv
streamlit
requests
